+++ back-end/Sacchon/src/main/java/gr/codehub/sacchon/app/repository/DoctorRepository.java
@@ -10,10 +10,10 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface DoctorRepository extends JpaRepository<Doctor, Integer>{
+public interface DoctorRepository extends JpaRepository<Doctor, Long>{
     @Query(value = "SELECT * FROM " + SacchonApplication.SCHEMA + ".DOCTOR WHERE email LIKE CONCAT('%' + ?1 + '%')", nativeQuery = true)
     List<Doctor> findDoctorByEmailNative(String match);
 
     @Query(value ="SELECT DOCTOR.FIRSTNAME, DOCTOR.LASTNAME, DOCTOR.EMAIL FROM " + SacchonApplication.SCHEMA + ".DOCTOR WHERE id =: id", nativeQuery = true)
-    Doctor findDoctorNameAndEmailById(@Param("id") int id);
+    Doctor findDoctorNameAndEmailById(@Param("id") long id);
 }
