+++ back-end/Sacchon/src/main/java/gr/codehub/sacchon/app/repository/GlucoseRecordRepository.java
@@ -10,9 +10,9 @@ import org.springframework.stereotype.Repository;
 import java.math.BigDecimal;
 
 @Repository
-public interface GlucoseRecordRepository extends JpaRepository<GlucoseRecord, Integer> {
+public interface GlucoseRecordRepository extends JpaRepository<GlucoseRecord, Long> {
     @Query(value = "SELECT AVG(GLUCOSE_RECORD.MEASUREMENT) FROM " + SacchonApplication.SCHEMA + ".GLUCOSE_RECORD " +
             "WHERE GLUCOSE_RECORD.GLUCOSE_ID = :glucoseId",
             nativeQuery = true)
-    BigDecimal findGlucoseRecordsByGlucoseId(@Param("glucoseId") int glucoseId);
+    BigDecimal findGlucoseRecordsByGlucoseId(@Param("glucoseId") long glucoseId);
 }
